pyproject.toml
[project]
name = "digital-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "flask-migrate>=4.0.5",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
    "requests>=2.32.5",
    "qrcode[pil]>=7.4.2",
]
